EWPP-436: Refactor the Node Wrapper classes.
