+++ modules/oe_content_call_proposals/tests/src/Kernel/CallForProposalsNodeWrapperTest.php
@@ -10,6 +10,7 @@ use Drupal\node\Entity\Node;
 use Drupal\node\Entity\NodeType;
 use Drupal\oe_content_call_proposals\CallForProposalsNodeWrapper;
 use Drupal\oe_content_call_proposals\CallForProposalsNodeWrapperInterface;
+use Drupal\oe_content\CallEntityWrapperInterface;
 
 /**
  * Tests "Call for proposals" wrapper class.
@@ -120,7 +121,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'hasDeadlineDate' => FALSE,
           'getOpeningDate' => NULL,
           'getDeadlineDate' => NULL,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_NOT_AVAILABLE,
+          'getStatus' => CallEntityWrapperInterface::STATUS_NOT_AVAILABLE,
           'hasStatus' => FALSE,
           'isUpcoming' => FALSE,
           'isOpen' => FALSE,
@@ -141,7 +142,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => TRUE,
           'isOpen' => FALSE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_UPCOMING,
+          'getStatus' => CallEntityWrapperInterface::STATUS_UPCOMING,
           'getStatusLabel' => 'Upcoming',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => FALSE,
@@ -165,7 +166,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => TRUE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_OPEN,
+          'getStatus' => CallEntityWrapperInterface::STATUS_OPEN,
           'getStatusLabel' => 'Open',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
@@ -189,7 +190,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => FALSE,
           'isClosed' => TRUE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_CLOSED,
+          'getStatus' => CallEntityWrapperInterface::STATUS_CLOSED,
           'getStatusLabel' => 'Closed',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
@@ -210,7 +211,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => FALSE,
           'isClosed' => TRUE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_CLOSED,
+          'getStatus' => CallEntityWrapperInterface::STATUS_CLOSED,
           'getStatusLabel' => 'Closed',
           'hasOpeningDate' => FALSE,
           'hasDeadlineDate' => TRUE,
@@ -238,7 +239,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => TRUE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_OPEN,
+          'getStatus' => CallEntityWrapperInterface::STATUS_OPEN,
           'getStatusLabel' => 'Open',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
@@ -272,7 +273,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => TRUE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_OPEN,
+          'getStatus' => CallEntityWrapperInterface::STATUS_OPEN,
           'getStatusLabel' => 'Open',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
@@ -306,7 +307,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => TRUE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_OPEN,
+          'getStatus' => CallEntityWrapperInterface::STATUS_OPEN,
           'getStatusLabel' => 'Open',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
@@ -337,7 +338,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => TRUE,
           'isOpen' => FALSE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_UPCOMING,
+          'getStatus' => CallEntityWrapperInterface::STATUS_UPCOMING,
           'getStatusLabel' => 'Upcoming',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => FALSE,
@@ -376,7 +377,7 @@ class CallForProposalsNodeWrapperTest extends CallForProposalsKernelTestBase {
           'isUpcoming' => FALSE,
           'isOpen' => TRUE,
           'isClosed' => FALSE,
-          'getStatus' => CallForProposalsNodeWrapperInterface::STATUS_OPEN,
+          'getStatus' => CallEntityWrapperInterface::STATUS_OPEN,
           'getStatusLabel' => 'Open',
           'hasOpeningDate' => TRUE,
           'hasDeadlineDate' => TRUE,
